+++ src/ol_openedx_events_handler/ol_openedx_events_handler/handlers/lms.py
@@ -3,8 +3,15 @@
 from ol_openedx_events_handler.handlers.course_access_role import (
     handle_course_access_role_added,
 )
+from ol_openedx_events_handler.handlers.enrollment import (
+    handle_course_enrollment_created,
+)
 from ol_openedx_events_handler.receivers.certificate_passing_receiver import (
     listen_for_passing_grade,
 )
 
-__all__ = ["handle_course_access_role_added", "listen_for_passing_grade"]
+__all__ = [
+    "handle_course_access_role_added",
+    "handle_course_enrollment_created",
+    "listen_for_passing_grade",
+]